Crate rid

Doc Template:

[short sentence explaining what it is]

[more detailed explanation]

[at least one code example that users can copy/paste to try it]

[even more advanced explanations if necessary]

Building

RID only supports running tests on the host.

‘cargo test –features=“std” <test_name>’

Modules

ptp

Precision Timing Protocol

rtnt

Real Time Numerical Tasks

Constants

RID_CYCLE_TIME_MS

Messaging rate for both host and client in milliseconds

RID_CYCLE_TIME_S

Messaging rate for both host and client in seconds

RID_CYCLE_TIME_US

Messaging rate for both host and client in microseconds

RID_DEFAULT_PID

Defualt Dyse Indstries product id

RID_DEFAULT_VID

Defualt Dyse Indstries vendor id

RID_MODE_INDEX

Mode of the packet

RID_PACKET_SIZE

HID packet size, tried going bigger and things broke

RID_PTP_RESERVED_BYTES

Bytes reserved for PTP

RID_TASK_INDEX

Id of the task this packet is meant for

Type Aliases

RIDReport

Simple type alias for more readability currently does not have any implementations