Module rid::ptp

Precision Timing Protocol

This crate can be included in a firmware build (use the client calls) or built using the “std” feature.

Structs

Duration

Lightweight timer object that runs for an hour, than resets.

TimeStamp

Precision Timing Protocol implementation

Constants

PTP_CRTS_INDEX

Client read time index, client sets and host reads this

PTP_CWTS_INDEX

Client write time index, client sets and host reads this

PTP_HRTS_INDEX

Host read time index, host sets and client reads this

PTP_HWTS_INDEX

Client write time index, host sets and client reads this

SEC_PER_HOUR

Second to Hour constant: seconds = hours * SEC_PER_HOUR

USEC_PER_SEC

Microsecond to Second constant: microseconds = seconds * USEC_PER_SEC

Functions

ptp_offset

Calculates the PTP offset