Module rid::rtnt

Real Time Numerical Tasks

Modules

constant

Real Time Task: Constant value

switch

Real Time Task: Switch

task_generator

Real Time Task Generator

task_manager

Real Time Task Manager

Structs

TaskConfig

Configuration data-structure for tasks.

TaskDataCache

A buffer for all system outputs

Enums

PacketType

Describes the contents of an RIDReport. These variants can control remote instances of the a task.

TaskStatus

Specifies the state a Task is in and the action required by the crate::rtnt::task_manager::TaskManager

Constants

MAX_TASKS

Maximum number of tasks user can spawn

MAX_TASK_CONFIG_CHUNKS

Maximum number of configuration chunks each task can use

MAX_TASK_DATA_BYTES

Maximum number of bytes in a Tasks output buffer and configuration chunk

MAX_TASK_DATA_FLOATS

Maximum nuber of floats in a tasks output (use f32, not worried about precision rn)

MAX_TASK_INPUTS

Maximum number of bytes a task can use in its name

RTNT_DATA_INDEX

RTNT Data start

RTNT_HDR_INDEX

Cnostant RTNT Header index

RTNT_HDR_LENGTH

RTNT Header length

Traits

RTNTask

Trait for RTNTasks

Type Aliases

InputIDBuffer

alias for readability

TaskBuffer

alias for readability

TaskData

alias for readability

TaskHeader

alias for readability