Struct rid::rtnt::TaskConfig

pub struct TaskConfig {
    id: u8,
    total_chunks: usize,
    missing_chunks: [bool; 32],
    buffer: [TaskBuffer; 32],
}

Configuration data-structure for tasks.

Tasks need more config data than fits into RID packets. This struct lets configurations get sharded into chunks and shared. The struct also notifies users when it has collected all the chunks.

Fields
id: u8

Tells the struct when a new set of chunks is coming

total_chunks: usizemissing_chunks: [bool; 32]

Missing chunks of the configuration data

buffer: [TaskBuffer; 32]

The data

Implementations

impl TaskConfig

pub fn default() -> TaskConfig

Create an empty configuration missing all chunks

pub fn new(total_chunks: usize, buffer: [TaskBuffer; 32]) -> TaskConfig

Create a buffer with the given number of chunks and buffer

pub fn chunks(&self) -> usize

Get the number of chunks

pub fn reset_chunks(&mut self)

Reset the chunks

Clears data. Does not change number of chunks

pub fn clear_missing(&mut self)

Clear the chunks

Clears the data and the number of chunks. This is required to make the buffer available for a new task.

pub fn init(&mut self)

Check if the buffer is being used

pub fn is_init(&self) -> bool

Check if the buffer is being used

pub fn clear_chunks(&mut self)

Clear the chunks

Clears the data and the number of chunks. This is required to make the buffer available for a new task.

pub fn new_chunk(&mut self, chunk_num: usize, chunk: &[u8])

Inserts chunk into data buffer and clear the missing flag

pub fn missing_chunks(&self) -> usize

Returns the number of missing chunks

This is only valid when total_chunks is set to the correct value.

pub fn first_missing(&self) -> Option<usize>

Returns the index of the first missing chunk

This is only valid when total_chunks is set to the correct value.

pub fn collect_chunk(&mut self, header: &[u8], buffer: &[u8])

Insert a new chunk or reset the current chunks

If the id changes it indicates a new configuration is available. When a new configuration becomes available the buffer will reset, then consume the chunk.

pub fn collect_status(&mut self, buffer: &[u8]) -> usize

Copy the chunk status from the buffer.

This will only happen on hosts trying to configure a set of tasks.

pub fn emit_chunk(&self) -> (TaskHeader, TaskBuffer)

Copy the configuration data to a buffer

Only happens on the host.

This assumes the missing chunks feild is syncronized with the client. That happens when the client sends a packet with the status.

pub fn emit_status(&self) -> TaskBuffer

Copy the status data to a buffer

Only happens on the client.

This synchronizes the missing data with the host. This should only happen when the client is configuring the task and the state is TaskStatus::Configuration

pub fn data(&self) -> &[TaskBuffer; 32]

Get a reference to the configuration data buffer