Struct rid::rtnt::TaskDataCache

pub struct TaskDataCache {
    context: usize,
    empty: TaskBuffer,
    buffer: [RIDReport; 32],
}

A buffer for all system outputs

Has a context that will set which task to stream. Can make this more complex if the field becomes a priority based queue, (i.e. stream = 0: no instances, stream = n: n uniformly distributed instances in queue) the queue can be built/restructured as nodes are created, each element of the queue will be the index of the desired task to stream. crate::rtnt::task_manager::TaskNode does not need to know if nodes stream.

Fields
context: usize

current task to publish

empty: TaskBuffer

Empty buffer for empty inputs

buffer: [RIDReport; 32]

buffer of task data

Implementations

impl TaskDataCache

pub fn default() -> TaskDataCache

Construct an empty cache

pub fn publish(&mut self, context_wrap: usize) -> Option<RIDReport>

Use the context to find the first task with unpublished output (bufferRID_TASK_INDEX > 0). Copy the output to a new buffer and clear the mode index.

pub fn status_waiting(&self, index: usize) -> bool

If the buffer has an unpublished PacketType::Status Used to prevent the client from overwriting status packets with data packets (Don’t stream data without telling the host the task is ready).

pub fn task_buffer(&mut self, index: usize) -> &[u8]

Mutable reference to a tasks buffer

pub fn new_output( &mut self, index: usize, status: PacketType, buffer: TaskBuffer )

Publish a new status and TaskBuffer to a tasks cache. Will always overwrite unpublished packets. If that is not desired The buffer should be checked before hand.

pub fn new_nonstreaming_output( &mut self, index: usize, mode: PacketType, buffer: TaskBuffer )

Write a new mode and TaskBuffer to a task’s cache. This packet can always be overwritten and will not stream to remote instances.

pub fn new_output_with_header( &mut self, index: usize, mode: PacketType, header: TaskHeader, buffer: TaskBuffer )

Publish a new PacketType, TaskHeader and TaskBuffer. This function will overwrite any existing data.

pub fn task_input_buffer(&self, inputs: &[u8; 16]) -> [&[u8]; 16]

Get a reference to all the requested task buffers. Not a mutable reference and does not make any garuantees about the data behind the reference.