Struct rid::rtnt::task_manager::TaskNode

pub struct TaskNode {
    pub rate: u16,
    pub stream: u8,
    pub n_inputs: u8,
    pub n_outputs: u8,
    pub inputs: InputIDBuffer,
    pub data: TaskData,
    pub status: TaskStatus,
    pub config_cache: TaskConfig,
    pub driver: Option<TaskDriver>,
    pub task: Option<TaskExecutable>,
}

Node containing an executable, stream, rate, inputs and status

Fields
rate: u16

Rate this task will execute at

stream: u8

Speceifies if the data should be streamed

n_inputs: u8

number of input values

n_outputs: u8

number of output values

inputs: InputIDBuffer

Input Tasks

data: TaskData

Output data

status: TaskStatus

Status of the task, enables and disables running the task

config_cache: TaskConfig

Buffer containing configuration data for the task

driver: Option<TaskDriver>

The driver assigned to this node

task: Option<TaskExecutable>

Optional Executable, there is always the max number of TaskNodes but not all will have TaskExecutables

Implementations

impl TaskNode

pub fn empty() -> TaskNode

Default constructor

pub fn new( stream: u8, rate: u16, inputs: u8, outputs: u8, driver: TaskDriver, task: TaskExecutable ) -> TaskNode

Create a TaskNode from a TaskExecutable and some header data.

This only happens on a host trying to configure TaskNodes from a file. On clients TaskNodes are built by collecting chunk packets from the host.

pub fn modify( &mut self, stream: u8, rate: u16, inputs: u8, outputs: u8, driver: TaskDriver, task: TaskExecutable )

Modify a TaskNode from a TaskExecutable and some header data.

This only happens on a host trying to configure TaskNodes from a file. On clients TaskNodes are built by collecting chunk packets from the host.

pub fn link(&mut self, link: InputIDBuffer)

Unnecesary setter, but sets the inputs

pub fn init(&mut self, header: &[u8], data: &[u8])

Copy the TaskNode header data from a buffer to Self

If the driver in the header changes the TaskNode will start configuring for that driver immediately. Called when a client recieves an init packet. Only clients should recieve init packets.

Also sets the TaskExecutable to the new driver (if it is a new driver).

pub fn header(&self) -> (TaskHeader, TaskBuffer)

Used to share the current header of a task. This is only useful on a host sending an init packet.

pub fn collect_chunk(&mut self, header: &[u8], data: &[u8])

Collect a packet containing config data. The data is only consumed if the TaskNode is already in the TaskStatus::Configuration state.

pub fn kill(&mut self)

Resets all task data and enters TaskStatus::Standby state.

pub fn configure(&mut self)

Configure the task. Activates the task if all chunks have been collected. If the TaskExecutable is not initialized there was an error initializing and the node reached this state in error, revert to standby.