Trait rid::rtnt::RTNTask

pub trait RTNTask {
    // Required methods
    fn default() -> Self;
    fn size(&self) -> usize;
    fn configure(&mut self, data: &[TaskBuffer]) -> bool;
    fn deconfigure(&self, data: &mut [TaskBuffer]) -> usize;
    fn run(&mut self, input: &[f32]) -> TaskData;
}

Trait for RTNTasks

Required Methods

fn default() -> Self

Create a new exectuable

fn size(&self) -> usize

The number of values in the output

fn configure(&mut self, data: &[TaskBuffer]) -> bool

Once the configuration chunks have all arrived pass the buffer to this. Given all chunks this function will initialize the RTNTasks members

fn deconfigure(&self, data: &mut [TaskBuffer]) -> usize

Convert the tasks members into a TaskConfig

fn run(&mut self, input: &[f32]) -> TaskData

Run the executable